[#50765485] Prompt for credentials when user clicks on Open File button on alpine work files with invalid credentials
Authorize the alpine workfile's data source.
Attempt to connect to the data source every time the alpine workfile is shown.

Also, use DataSourceCredentials mixin in Alpine workfile and work flow show page
and ensure that the user is prompted for credentials only if the data source
authorization fails (as opposed to the workspace authorization).