[#44827221] Support metadata queries for HAWQ
The 'pg_total_relation_size' function does not exist in HAWQ.
The only way we could think of in the meantime to differentiate between HAWQ and GPDB
was to actually run the query and if it fails, then try the HAWQ variant.

The HAWQ variant includes a separate query to calculate the disk size.