backboneModelEqualityTester is now more likely to return true/false to prevent pathological object comparisons.
Previously, it only returned 'true' when comparing objects by cid. Now,
if two objects have a cid and they are not equal, it returns 'false'.

This prevents Jasmine from carrying on comparing two objects when we
already know they're not similar as far as Backbone is concerned. This
occasionally added two seconds to chorus.models.ChorusView spec, which
did some 'toContain()' matching with models.