[#fixes 57181288] Reordering Job Tasks is more reliable
Previously, clicking on the same arrow repeatedly could result in two tasks
with the same index. Now that is less likely.

Job tasks are now reordered by PUTting the list of task ids to the Job update endpoint.

Job tasks are now a sub-collection of the job model, and should fire 'reset'
whenever the job is loaded or saved.