[Finishes #53587977] note.rb can probably be safely reloaded during development now
Previously, there was some weird cruft at the bottom of note.rb which made it so notes
could not be created during development if note.rb was every reloaded by
Rails (like, if you ever added or changed a method in it or its related
classes).

We removed the cruft and found a bug with AttachmentAccess, which is the only thing
we saw fail immediately. There may be more land mines in this area.